+++ app/src/main/AndroidManifest.xml
@@ -11,15 +11,15 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.ExhibitionDoT"
+        android:name=".DoTApplication"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name="com.exhibitiondot.presentation.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.ExhibitionDoT">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
