+++ app/src/main/java/com/exhibitiondot/android/DoTApplication.kt
@@ -1,0 +1,7 @@
+package com.exhibitiondot.android
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class DoTApplication : Application()
