+++ presentation/src/main/java/com/exhibitiondot/presentation/MainActivity.kt
@@ -1,0 +1,18 @@
+package com.exhibitiondot.presentation
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        enableEdgeToEdge()
+        super.onCreate(savedInstanceState)
+
+        setContent {
+
+        }
+    }
+}
